E-Checkout_BE/requirements.txt
telnyx~=1.0.0b4
Django==2.1
django-cors-headers==3.1.0
djangorestframework~=3.9.4
djangorestframework-jwt==1.11.0
psycopg2==2.7.7
PyJWT~=1.7.0
pytz==2019.3
xlwt==1.3.0

requests~=2.21.0
docusign-esign==3.2.0